Switch to `docker compose` (aka V2) from `docker-compose` (aka V1)
This will harmonize the container name scheme separator to '-'.
This matters because Docker Desktop has a setting that automatically
converts V1 calls to V2 calls, which can be problematic if we use
commands relying on container names.

We also switch to using `docker wait` instead of `ps` to get exit code.